+++ R/register_all_transpilers.R
@@ -39,6 +39,7 @@ register_all_transpilers <- function() {
   transpilers_for_package("futurize::add-on", package = "kernelshap",     append_transpilers_for_kernelshap)
   transpilers_for_package("futurize::add-on", package = "glmmTMB",        append_transpilers_for_glmmTMB)
   transpilers_for_package("futurize::add-on", package = "lme4",           append_transpilers_for_lme4)
+  transpilers_for_package("futurize::add-on", package = "metafor",        append_transpilers_for_metafor)
   transpilers_for_package("futurize::add-on", package = "partykit",       append_transpilers_for_partykit)
   transpilers_for_package("futurize::add-on", package = "riskRegression", append_transpilers_for_riskRegression)
   transpilers_for_package("futurize::add-on", package = "seriation",      append_transpilers_for_seriation)