+++ NEWS.md
@@ -42,6 +42,9 @@
  * Add support for domain-specific Bioconductor package **GSVA**,
    e.g. `es <- gsva(gsvaParam(expr, geneSets)) |> futurize()`.
 
+ * Add support for domain-specific CRAN package **metafor**, e.g. `pr
+   <- profile(fit) |> futurize()`.
+
  * Add support for domain-specific CRAN package **partykit**, e.g. `cf
    <- cforest(dist ~ speed, data = cars) |> futurize()`.